Renamed all files to snake case
