+++ posenet/demos/index.html
@@ -8,8 +8,8 @@
     <ul>
       <li><a href='camera.html'>Pose Estimation - Camera feed demo</a></li>
       <li><a href='coco.html'>Pose Estimation - Image demo</a></li>
-      <li><a href='partMapCamera.html'>Segmentation & Part Map Estimation - Camera Demo</a></li>
-      <li><a href='partMapImage.html'>Segmentation & Part Map Estimation - Image demo</a></li>
+      <li><a href='part_map_camera.html'>Segmentation & Part Map Estimation - Camera Demo</a></li>
+      <li><a href='part_map_image.html'>Segmentation & Part Map Estimation - Image demo</a></li>
     </ul>
   </body>
 </html>
