+++ posenet/demos/part_map_camera.html
@@ -53,7 +53,7 @@
             </p>
         </div>
     </div>
-    <script src="partMapCamera.js"></script>
+    <script src="part_map_camera.js"></script>
 </body>
 
 </html>
