+++ posenet/demos/part_map_camera.js
