+++ posenet/demos/part_map_image.html
@@ -96,7 +96,7 @@
       </p>
     </div>
   </div>
-  <script src="partMapImage.js"></script>
+  <script src="part_map_image.js"></script>
 </body>
 
 </html>
