+++ posenet/src/mobilenet.ts
@@ -16,7 +16,7 @@
  */
 
 import * as tf from '@tensorflow/tfjs';
-import {ModelWeights} from './modelWeights';
+import {ModelWeights} from './model_weights';
 
 export type MobileNetMultiplier = 0.25|0.50|0.75|1.0|1.01;
 export type ConvolutionType = 'conv2d'|'separableConv';
