+++ posenet/src/multi_pose/util.ts
