+++ posenet/src/single_pose/argmax2d.ts
